Fix unit of measurement for charge added shown in overview, documentation to be correct (#1061)
* Fix units of charge energy added to be correct

The name of the measurement didn't make sense, since it is plotting added energy but was using an unit of power (kW). Changed to make sense.

* Fix unit of energy added

Documentation was using the wrong unit for energy added, fixed.

* Fix unit of energy added

Documentation was using unit of power (kW) for measuring energy added. Fixed to use the correct unit (kWh)